Fix console display issues with seeds
"Please select a seed" is now displayed properly when the player must choose a seed to use.

Some values were casted to ints when being printed to the console to avoid leading 0s on them.

"Press X/Y to select a seed" is no longer displayed at boot since this is redundant.

+++ GameCube/subrel/boot/source/rando/seedlist.cpp
@@ -80,7 +80,7 @@ namespace mod::rando
             }
         }
 
-        mod::console << m_numSeeds << " seed(s) available.\n";
+        mod::console << static_cast<int32_t>( m_numSeeds ) << " seed(s) available.\n";
 
         delete[] headerBuffer;
     }
